- During testin on 64 bit linux GCC 4.3.3 found a cast which   was not explicit in calculating menu position and corrected.
Testing
- Compile Linux 64 bit GCC 4.3.3 X11
- Several interactive sessions.




Former-commit-id: fa2ba36badbc4676b98c0d717ec138fdf0a4eb96

+++ src/lib/esru_x.c
@@ -7131,7 +7131,7 @@ int		len_title;
              XSetForeground(theDisp,theGC, fg); XSetBackground(theDisp,theGC, bg);
              Timer(300);
              xbox(menubx,fg,white,BMEDGES|BMCLEAR);	/* clear menu box */
-             *ino = (long int) i +1;  /* compensate for c starting at zero */
+             *ino = (long int) i + (long int) 1;  /* compensate for c starting at zero */
              *uresp = (long int) 0;
    	     break;
            }
